+++ app/src/main/java/com/numero/material_gallery/components/list/ComponentListActivity.kt
@@ -22,6 +22,7 @@ import com.numero.material_gallery.components.button.MaterialButtonActivity
 import com.numero.material_gallery.components.button.MaterialButtonToggleGroupActivity
 import com.numero.material_gallery.components.card.MaterialCardActivity
 import com.numero.material_gallery.components.chip.ChipActivity
+import com.numero.material_gallery.components.dialog.MaterialDialogActivity
 import com.numero.material_gallery.components.fab.ExtendedFabActivity
 import com.numero.material_gallery.components.fab.FabActivity
 import com.numero.material_gallery.components.image.ShapeableImageViewActivity
