+++ app/src/main/res/layout/activity_material_dialog.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.MaterialDialogActivity">
+    tools:context=".components.dialog.MaterialDialogActivity">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
